add label for output options
