+++ application/ui/src/features/inspect/toolbar/sinks/mqtt-fields copy/mqtt-fields.component.tsx
