+++ application/ui/src/features/inspect/toolbar/sinks/output-formats/output-formats.module.scss
@@ -1,0 +1,7 @@
+.itemList {
+    [role='group'] {
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+    }
+}
